Corrected serious bug in world handling
This one slipped in after some late night refactoring. head used to get
passed into read_name_value were it would be increased, since this was
pointless I changed it so head was always increased at the end of the
loop. I forgot to remove the code increasing from all the other handlers
in the if block though. Awkward.