Added ucfb_reader abstraction
So it begins...

The goal is to replace all the "raw" access to the chunks with a single
unified, safer, easily followed abstraction. Importantly it'll add
safety with close to zero-overhead and will not impede the ability to
parallelize the processing. Clean code, safety, very low overhead,
parallelism, all the things I love.