Isolated msh::Builders_map for each parent chunk
This fixes cases where individual lvl_ chunks each had a copy of a
model, resulting in the models being merged together.

Now whenever we find a "parent" chunk we create a new msh::Builders_map
object for it's children to use instead. This makes sure the models are
not merged together, however we may end up writing the same model out
twice in same cases. I would expect that to have a negligible impact on
overall app performance however.