Changed the way config file numbers are handled
Previously they were always written out as floats, because that is how
they are stored and treated by the munger. This caused problems however
when Zero Editor tried to read certain values as integers, leading to it
crashing. Now we take an educated guess as to if a number should be
written out as an int or float based off size of it's fractional part.